+++ apps/core/src/components/providers/index.ts
@@ -1,6 +1,5 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './QR';
-
-export * from './providers';
+export * from './KioskClientProvider';
+export * from './ThemeProvider';
