+++ apps/core/src/enums/index.ts
@@ -1,6 +1,4 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './QR';
-
-export * from './providers';
+export * from './theme.enums';
