+++ apps/core/src/enums/theme.enums.ts
@@ -1,6 +1,7 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './QR';
-
-export * from './providers';
+export enum Theme {
+    Light = 'light',
+    Dark = 'dark',
+}
